add fix to screen header to be hidden on bottom navigation
