+++ __test__/components/card/index.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 import { fireEvent, render } from '@testing-library/react-native';
-import Theme from '../../../theme';
-import Card from '../../../components/card';
+import Theme from '../../../src/theme';
+import Card from '../../../src/components/card';
 
 describe('TEST CARD COMPONENT(<Card/>)', () => {
   const onPress = jest.fn();
@@ -24,7 +24,7 @@ describe('TEST CARD COMPONENT(<Card/>)', () => {
 
   test('It renders other components passed as children to the card <Card/> component accurately', () => {
     const cardContainer = getByTestId('card-container');
-    expect(cardContainer).toBeTruthy();
+    expect(cardContainer.children).toBeTruthy();
   });
 
   test('card component on render can be clicked', () => {
