+++ __test__/config/jest.setup.ts
@@ -1,0 +1,5 @@
+import { cleanup } from '@testing-library/react-native';
+
+afterEach(cleanup);
+
+afterAll(cleanup);
