+++ __test__/screens/signin/index.test.tsx
@@ -1,26 +1,20 @@
 import React from 'react';
 import { fireEvent, render } from '@testing-library/react-native';
-import Theme from '../../../theme';
-import { createNavigationTestProps } from '../../../constants';
-import SignUpScreen from '../../../screens/signup';
+import Theme from '../../../src/theme';
+import { createNavigationTestProps } from '../../../src/constants';
+import SignInScreen from '../../../src/screens/signin';
 
 const mountComponent = () => {
-  const onHandleChange = jest.fn(() => {});
-
   const props: any = createNavigationTestProps();
   const renderedProps = render(
     <Theme>
-      <SignUpScreen {...props} />
+      <SignInScreen {...props} />
     </Theme>
   );
-  return {
-    props,
-    ...renderedProps,
-    onHandleChange
-  };
+  return { props, ...renderedProps };
 };
 
-describe(' TEST GET_SIGNUP SCREEN (<SignUpScreen/>)', () => {
+describe(' TEST SIGN_IN SCREEN (<SignInScreen/>)', () => {
   test('It renders the App logo component', () => {
     const { queryByTestId } = mountComponent();
     expect(queryByTestId('appLogo')).toBeTruthy();
@@ -28,10 +22,19 @@ describe(' TEST GET_SIGNUP SCREEN (<SignUpScreen/>)', () => {
 
   test('That the screen has all the necessary input fields', () => {
     const { getByTestId } = mountComponent();
-    expect(getByTestId('userName')).toBeTruthy();
     expect(getByTestId('email')).toBeTruthy();
-    expect(getByTestId('phone')).toBeTruthy();
     expect(getByTestId('password')).toBeTruthy();
+  });
+
+  test('That the forgot password button was pressed', () => {
+    const { props, queryByTestId } = mountComponent();
+    const signUpButton = queryByTestId('resetAccount');
+    expect(signUpButton).toBeTruthy();
+    fireEvent.press(queryByTestId('resetAccount'));
+    expect(props.navigation.navigate).toHaveBeenCalledTimes(1);
+    expect(props.navigation.navigate).toHaveBeenCalledWith(
+      'ForgotPasswordScreen'
+    );
   });
 
   test('That the login button was pressed', () => {
@@ -39,18 +42,16 @@ describe(' TEST GET_SIGNUP SCREEN (<SignUpScreen/>)', () => {
     const loginButton = queryByTestId('loginButton');
     expect(loginButton).toBeTruthy();
     fireEvent.press(queryByTestId('loginButton'));
-    expect(props.navigation.navigate).toHaveBeenCalledTimes(1);
-    expect(props.navigation.navigate).toHaveBeenCalledWith('SignInScreen');
+    expect(props.navigation.replace).toHaveBeenCalledTimes(1);
+    expect(props.navigation.replace).toHaveBeenCalledWith('HomeScreen');
   });
 
   test('That the sigUp button was pressed', () => {
     const { props, queryByTestId } = mountComponent();
-    const signUpButton = queryByTestId('submitButton');
+    const signUpButton = queryByTestId('createAccount');
     expect(signUpButton).toBeTruthy();
-    fireEvent.press(queryByTestId('submitButton'));
+    fireEvent.press(queryByTestId('createAccount'));
     expect(props.navigation.navigate).toHaveBeenCalledTimes(1);
-    expect(props.navigation.navigate).toHaveBeenCalledWith(
-      'ProfileSetupScreen'
-    );
+    expect(props.navigation.navigate).toHaveBeenCalledWith('SignUpScreen');
   });
 });
