+++ __test__/utils/applyScale..test.ts
@@ -1,7 +1,7 @@
-import applyScale from '../../utils/applyScale';
-import Devices from '../../lib/responsiveImage/device';
+import applyScale from '../../src/utils/applyScale';
+import Devices from '../../src/libs/responsiveImage/device';
 
-jest.mock('Dimensions');
+// jest.mock('Dimensions');
 
 describe('TEST IMAGE SCALING FUNCTIONALITY', () => {
   test('get device screen size and return scaled value', () => {
