+++ src/libs/dummyPost.json
@@ -1,0 +1,31 @@
+{
+  "images": [],
+  "contentType": "community",
+  "link": null,
+  "category": {
+    "_id": "5d5e640247fda3145df0d912",
+    "title": "First trimester",
+    "id": "5d5e640247fda3145df0d912"
+  },
+  "isFlagged": false,
+  "deleted": false,
+  "_id": "5de102cc7fcd6d248dbbf4ee",
+  "topic": "Is it normal for my baby to move this way at 6 months?",
+  "description": "As conscious traveling Paupers we must always be concerned about our dear Mother Earth. If you think about it, you travel across her face.",
+  "user": {
+    "avatar": "https://babybliss-platform.sfo2.digitaloceanspaces.com/avatar/avatar-5fcae2-H1pRhXVAkUuJ0NunOgXAhKDdl-1567081512239.png",
+    "_id": "5d5fcae2848a873ccd4f1c58",
+    "name": "BabyBliss Moderator",
+    "username": "moderator",
+    "id": "5d5fcae2848a873ccd4f1c58"
+  },
+  "createdAt": "2019-11-29T11:36:44.699Z",
+  "updatedAt": "2019-11-29T11:36:48.138Z",
+  "__v": 1,
+  "id": "5de102cc7fcd6d248dbbf4ee",
+  "noOfComments": 0,
+  "isLiked": false,
+  "noOfLikes": 900,
+  "noOfViews": 130,
+  "userId": "5d5fcae2848a873ccd4f1c58"
+}
