+++ src/libs/responsiveImage/styles.ts
@@ -1,10 +1,5 @@
 import styled from 'styled-components/native';
 
-export const Container = styled.View`
-  justify-content: center;
-  align-items: center;
-`;
-
 export const Image = styled.Image`
   position: absolute;
   top: 0;
