+++ src/utils/applyScale.ts
@@ -1,4 +1,4 @@
-import Device from '../lib/responsiveImage/device';
+import Device from '../libs/responsiveImage/device';
 
 export default function applyScale(size: number) {
   return Math.ceil(size * Device.scale);
